api/requirements.txt
azure-functions
pandas
openpyxl
pyrfc
azure-data-tables
cryptography
